目前完成的第一代正式版本，模型是流式输出，链接了几个rag向量数据库

+++ xiaoY/src/main/java/com/pj/xiaoY/entity/CourseComment.java
@@ -29,7 +29,7 @@ public class CourseComment implements Serializable {
 	/**
 	 * 关联课程ID，对应course表的id
 	 */
-	private Long courseId;
+	private Integer courseId;
 	/**
 	 * 评论用户ID
 	 */
